refactor(core): move wait reducers to scoped packages
